Lucas service controller

+++ backend/src/main/java/ca/mcgill/ecse321/Mar1HotelSystem/model/Manager.java
@@ -19,7 +19,7 @@ public class Manager extends Account {
 
     // Manager constructor requiring firstName, lastName, email, phoneNUmber,
     // password
-    public Manager(String firstName, String lastName, String email, int phoneNumber, String password) {
+    public Manager(String firstName, String lastName, String email, long phoneNumber, String password) {
         super(firstName, lastName, email, phoneNumber, password);
     }
